backend/requirements.txt
Flask
Flask-Cors
requests
python-dotenv
pandas
scikit-learn
gunicorn
